refactor: move keyboard query to new directory
